+++ presentation/src/main/java/com/meongmoryteam/presentation/ui/register_family/RegisterFamilyScreen.kt
@@ -2,26 +2,16 @@ package com.meongmoryteam.presentation.ui.register_family
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.meongmoryteam.presentation.ui.theme.ButtonContent
 import com.meongmoryteam.presentation.ui.theme.Typography
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.style.TextAlign
 import androidx.navigation.NavController
 import com.meongmoryteam.presentation.R
-
 import com.meongmoryteam.presentation.ui.theme.Brown
 import com.meongmoryteam.presentation.ui.theme.DarkGrey
 import com.meongmoryteam.presentation.ui.theme.Orange
@@ -59,34 +49,5 @@ fun RegisterFamilyScreen(navController: NavController) {
                 colors = ButtonDefaults.textButtonColors(containerColor = Brown, contentColor = ButtonContent),
                 style = Typography.labelMedium)
         }
-    }
-}
-
-@Composable
-fun TextComponent(text: String, style: TextStyle, modifier: Modifier = Modifier, color: Color){
-    Text(
-        text = text,
-        style = style,
-        modifier = modifier,
-        color = color)
-}
-
-@Composable
-fun TextButtonComponent(onClick: ()->Unit, text: String, colors: ButtonColors, style: TextStyle, width: Float = 1f){
-    TextButton(
-        onClick = onClick,
-        shape = RoundedCornerShape(10.dp),
-        modifier= Modifier
-            .fillMaxWidth(width)
-            .height(50.dp)
-            .padding(vertical = 5.dp),
-        colors = colors,
-    )
-    {
-        Text(
-            text = text,
-            style = style,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.padding(vertical = 0.dp))
     }
 }
